requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# Show-Attend-Tell/data.py: 3
# Show-Attend-Tell/eval.py: 7
Pillow == 7.0.0

# Show-Attend-Tell/train.py: 13,14
apex == 0.1

# Show-Attend-Tell/data.py: 1
h5py == 2.10.0

# Show-Attend-Tell/eval.py: 9,10
# Show-Attend-Tell/train.py: 9
matplotlib == 3.1.1

# Show-Attend-Tell/commons.py: 2
# Show-Attend-Tell/eval.py: 5,6
# Show-Attend-Tell/train.py: 15,16
nltk == 3.4.5

# Show-Attend-Tell/data.py: 4
# Show-Attend-Tell/eval.py: 11
# Show-Attend-Tell/train.py: 11
numpy == 1.18.1

# Show-Attend-Tell/main.py: 6
# Show-Attend-Tell/train.py: 10
pandas == 1.0.1

# Show-Attend-Tell/eval.py: 12
scikit_image == 0.16.2

# Show-Attend-Tell/commons.py: 1
# Show-Attend-Tell/data.py: 5,6
# Show-Attend-Tell/eval.py: 2
# Show-Attend-Tell/main.py: 4,5
# Show-Attend-Tell/models.py: 1,2
# Show-Attend-Tell/train.py: 2,3,5
torch == 1.4.0

# Show-Attend-Tell/eval.py: 3
# Show-Attend-Tell/main.py: 9
# Show-Attend-Tell/models.py: 3
# Show-Attend-Tell/train.py: 6
torchvision == 0.5.0

# Show-Attend-Tell/data.py: 1
# Show-Attend-Tell/eval.py: 4
# Show-Attend-Tell/models.py: 5
# Show-Attend-Tell/train.py: 8
tqdm == 4.43.0
